refactor: flatten folder structure and fix embedding pipeline bugs
Structural simplification (57 files/19 dirs → 45 files/10 dirs):
- Eliminate all single-file folders and empty pass-through directories
- Flatten ui/embedding/, ui/models/embed/adapters/, domain/models/embed/
- Merge utils (7→1), domain config files (3→1), embed-model (3→1), kernel (6→2)
- Remove dead code (furthest_acc, dead barrel exports)

Bug fixes:
- Fix _chunked_pipeline_active stuck on exception (try/finally)
- Preserve re-import paths across embed_job_queue.clear()
- Trigger post-chunked debounceReImport for paths queued during processing
- Fix removeBySourcePath memory leak (stale insertionOrder keys)
- Fix silent error swallowing in codeblock processor and ConnectionsView
- Move deferRetryCount to plugin instance (was module-level singleton)
- Inline 7 single-use helper functions in embed-orchestrator
