+++ src/domain/embedding-pipeline.ts
