--- src/domain/embedding/kernel/reducer.ts
@@ -1,194 +1,0 @@
-/**
- * @file embedding/kernel/reducer.ts
- * @description Pure reducer for embedding kernel state transitions (3-state: idle/running/error)
- */
-
-import type {
-  EmbeddingKernelEvent,
-  EmbeddingKernelState,
-} from './types';
-
-export function createInitialKernelState(): EmbeddingKernelState {
-  return {
-    phase: 'idle',
-    model: null,
-    run: null,
-    queue: {
-      pendingJobs: 0,
-      staleTotal: 0,
-      staleEmbeddableTotal: 0,
-      queuedTotal: 0,
-    },
-    lastError: null,
-  };
-}
-
-export function reduceEmbeddingKernelState(
-  prev: EmbeddingKernelState,
-  event: EmbeddingKernelEvent,
-): EmbeddingKernelState {
-  switch (event.type) {
-    case 'INIT_CORE_READY':
-      return prev;
-
-    case 'INIT_CORE_FAILED':
-      return {
-        ...prev,
-        phase: 'error',
-        lastError: {
-          code: 'INIT_CORE_FAILED',
-          message: event.error,
-          at: Date.now(),
-        },
-      };
-
-    case 'MODEL_SWITCH_REQUESTED':
-      return {
-        ...prev,
-        phase: prev.phase === 'error' ? 'idle' : prev.phase,
-        lastError: null,
-      };
-
-    case 'MODEL_SWITCH_SUCCEEDED':
-      return {
-        ...prev,
-        phase: 'idle',
-        model: event.model,
-        run: null,
-        lastError: null,
-      };
-
-    case 'MODEL_SWITCH_FAILED':
-      return {
-        ...prev,
-        phase: 'error',
-        run: null,
-        lastError: {
-          code: 'MODEL_SWITCH_FAILED',
-          message: event.error,
-          context: event.reason,
-          at: Date.now(),
-        },
-      };
-
-    case 'QUEUE_SNAPSHOT_UPDATED':
-      return {
-        ...prev,
-        queue: event.queue,
-      };
-
-    case 'QUEUE_HAS_ITEMS':
-      if (prev.phase !== 'idle') return prev;
-      return {
-        ...prev,
-        phase: 'running',
-      };
-
-    case 'QUEUE_EMPTY':
-      if (prev.phase !== 'running') return prev;
-      return {
-        ...prev,
-        phase: 'idle',
-        run: null,
-      };
-
-    case 'RUN_REQUESTED':
-      return {
-        ...prev,
-        lastError: null,
-      };
-
-    case 'RUN_STARTED': {
-      if (prev.phase !== 'idle') {
-        console.warn(`[SC][FSM] RUN_STARTED blocked: cannot start from '${prev.phase}'`);
-        return prev;
-      }
-      return {
-        ...prev,
-        phase: 'running',
-        run: event.run,
-        lastError: null,
-      };
-    }
-
-    case 'RUN_PROGRESS':
-      if (!prev.run) return prev;
-      return {
-        ...prev,
-        run: {
-          ...prev.run,
-          current: event.current,
-          total: event.total,
-          currentEntityKey:
-            event.currentEntityKey === undefined
-              ? prev.run.currentEntityKey
-              : event.currentEntityKey,
-          currentSourcePath:
-            event.currentSourcePath === undefined
-              ? prev.run.currentSourcePath
-              : event.currentSourcePath,
-        },
-      };
-
-    case 'RUN_FINISHED':
-      return {
-        ...prev,
-        phase: 'idle',
-        run: null,
-      };
-
-    case 'RUN_FAILED':
-      return {
-        ...prev,
-        phase: 'error',
-        run: null,
-        lastError: {
-          code: 'RUN_FAILED',
-          message: event.error,
-          at: Date.now(),
-        },
-      };
-
-    case 'FATAL_ERROR':
-      if (prev.phase !== 'running') return prev;
-      return {
-        ...prev,
-        phase: 'error',
-        run: null,
-        lastError: {
-          code: event.code,
-          message: event.error,
-          at: Date.now(),
-        },
-      };
-
-    case 'REFRESH_REQUESTED':
-    case 'REIMPORT_REQUESTED':
-      return {
-        ...prev,
-        lastError: null,
-      };
-
-    case 'REIMPORT_COMPLETED':
-      return prev;
-
-    case 'REIMPORT_FAILED':
-      return {
-        ...prev,
-        lastError: {
-          code: 'REIMPORT_FAILED',
-          message: event.error,
-          at: Date.now(),
-        },
-      };
-
-    case 'RESET_ERROR':
-      return {
-        ...prev,
-        lastError: null,
-      };
-
-    default:
-      return prev;
-  }
-}
