--- src/domain/search/index.ts
@@ -1,9 +1,0 @@
-/**
- * @file index.ts
- * @description Barrel exports for search module
- */
-
-export {
-  EmbeddingPipeline,
-  type EmbedQueueStats,
-} from './embedding-pipeline';
