--- src/types/index.ts
@@ -1,34 +1,0 @@
-/**
- * @file index.ts
- * @description Barrel exports for shared types
- * Import from 'src/types' to get all type definitions
- */
-
-// Settings types
-export type {
-  EmbedModelSettings,
-  SourceSettings,
-  BlockSettings,
-  ViewFilterSettings,
-  PluginSettings,
-} from './settings';
-
-// Entity types
-export type {
-  LastMetadata,
-  EmbeddingData,
-  EntityData,
-  SourceData,
-  BlockData,
-  ConnectionResult,
-  EmbeddingEntity,
-  SearchFilter,
-} from './entities';
-
-// Model types
-export type {
-  ModelInfo,
-  EmbedInput,
-  EmbedResult,
-  EmbedModelAdapter,
-} from './models';
