+++ src/ui/embed-adapters/gemini.ts
@@ -8,8 +8,8 @@ import {
   EmbedModelRequestAdapter,
   EmbedModelResponseAdapter,
 } from './api-base';
-import type { EmbedInput, EmbedResult, ModelInfo } from '../../../../types/models';
-import { embedAdapterRegistry } from '../../../../domain/models/embed/registry';
+import type { EmbedInput, EmbedResult, ModelInfo } from '../../types/models';
+import { embedAdapterRegistry } from '../../domain/embed-model';
 
 export const GEMINI_SIGNUP_URL = 'https://aistudio.google.com/apikey';
