+++ src/ui/file-watcher.ts
@@ -61,11 +61,9 @@ export function isSourceFile(file: TFile): boolean {
 }
 
 export function queueSourceReImport(plugin: SmartConnectionsPlugin, path: string): void {
-  // Enqueue into the unified EmbedJobQueue. LWW dedup ensures rapid edits
-  // to the same file collapse into one entry with the latest timestamp.
   plugin.embed_job_queue?.enqueue({
     entityKey: path,
-    contentHash: '', // unknown until file is imported
+    contentHash: '',
     sourcePath: path.split('#')[0],
     enqueuedAt: Date.now(),
   });
@@ -73,19 +71,14 @@ export function queueSourceReImport(plugin: SmartConnectionsPlugin, path: string
 }
 
 export function removeSource(plugin: SmartConnectionsPlugin, path: string): void {
-  // Remove from the unified queue (source + all block keys under this source)
   plugin.embed_job_queue?.removeBySourcePath(path);
-
-  if (plugin.source_collection) {
-    plugin.source_collection.delete(path);
-  }
-
-  if (plugin.block_collection) {
-    plugin.block_collection.delete_source_blocks(path);
-  }
+  plugin.source_collection?.delete(path);
+  plugin.block_collection?.delete_source_blocks(path);
 }
 
 export function debounceReImport(plugin: SmartConnectionsPlugin): void {
+  if (plugin._chunked_pipeline_active) return;
+
   plugin.re_import_halted = true;
   if (plugin.re_import_timeout) {
     window.clearTimeout(plugin.re_import_timeout);
@@ -97,7 +90,7 @@ export function debounceReImport(plugin: SmartConnectionsPlugin): void {
 
   const waitTime = (plugin.settings.re_import_wait_time || 13) * 1000;
   plugin.re_import_timeout = window.setTimeout(() => {
-    void enqueueReImportJob(plugin, 'Debounced re-import').catch((error) => {
+    void enqueueReImportJob(plugin).catch((error) => {
       console.error('Failed to enqueue debounced re-import:', error);
     });
   }, waitTime);
@@ -106,32 +99,27 @@ export function debounceReImport(plugin: SmartConnectionsPlugin): void {
 }
 
 const MAX_DEFER_RETRIES = 20;
-let deferRetryCount = 0;
 
 function deferReImport(plugin: SmartConnectionsPlugin, reason: string, delayMs: number = 1500): void {
-  deferRetryCount++;
-  if (deferRetryCount > MAX_DEFER_RETRIES) {
-    console.warn(`[SC] Re-import deferred ${deferRetryCount} times — giving up. Reason: ${reason}`);
-    deferRetryCount = 0;
+  plugin._defer_retry_count++;
+  if (plugin._defer_retry_count > MAX_DEFER_RETRIES) {
+    console.warn(`[SC] Re-import deferred ${plugin._defer_retry_count} times — giving up. Reason: ${reason}`);
+    plugin._defer_retry_count = 0;
     return;
   }
-  console.log(`${reason}. Deferring re-import for ${delayMs}ms (attempt ${deferRetryCount}/${MAX_DEFER_RETRIES})...`);
+  console.log(`${reason}. Deferring re-import for ${delayMs}ms (attempt ${plugin._defer_retry_count}/${MAX_DEFER_RETRIES})...`);
   if (plugin.re_import_retry_timeout) {
     window.clearTimeout(plugin.re_import_retry_timeout);
   }
   plugin.re_import_retry_timeout = window.setTimeout(() => {
     plugin.re_import_retry_timeout = undefined;
-    void enqueueReImportJob(plugin, reason).catch((error) => {
+    void enqueueReImportJob(plugin).catch((error) => {
       console.error('Failed to enqueue deferred re-import:', error);
     });
   }, delayMs);
 }
 
-function resetDeferRetryCount(): void {
-  deferRetryCount = 0;
-}
-
-function enqueueReImportJob(plugin: SmartConnectionsPlugin, _reason: string): Promise<void> {
+function enqueueReImportJob(plugin: SmartConnectionsPlugin): Promise<void> {
   return plugin.enqueueEmbeddingJob({
     type: 'REIMPORT_SOURCES',
     key: 'REIMPORT_SOURCES',
@@ -140,10 +128,6 @@ function enqueueReImportJob(plugin: SmartConnectionsPlugin, _reason: string): Pr
   });
 }
 
-/**
- * Collect source-level file paths from the EmbedJobQueue.
- * Only returns entries whose entityKey looks like a source path (no '#' block ref).
- */
 function getReImportPaths(plugin: SmartConnectionsPlugin): string[] {
   if (!plugin.embed_job_queue) return [];
   return plugin.embed_job_queue
@@ -158,6 +142,11 @@ export async function runReImport(plugin: SmartConnectionsPlugin): Promise<void>
 
   if (!plugin.source_collection || !plugin.embedding_pipeline) {
     console.warn('Collections or pipeline not initialized');
+    return;
+  }
+
+  if (plugin._chunked_pipeline_active) {
+    console.log('[SC] Re-import skipped: chunked pipeline is active, paths remain queued');
     return;
   }
 
@@ -202,22 +191,19 @@ export async function runReImport(plugin: SmartConnectionsPlugin): Promise<void>
 
     await plugin.runEmbeddingJobImmediate(`Re-import (${queue_paths.length} files)`);
 
-    // Remove processed source-level entries from the queue
     for (const path of processed_paths) {
       plugin.embed_job_queue?.removeBySourcePath(path);
     }
 
     plugin.refreshStatus();
-    resetDeferRetryCount();
+    plugin._defer_retry_count = 0;
     console.log('Re-import completed');
     plugin.dispatchKernelEvent({ type: 'REIMPORT_COMPLETED' });
 
-    // If the queue still has source-level items, schedule another re-import
     if (!plugin._unloading && getReImportPaths(plugin).length > 0) {
       deferReImport(plugin, 'Re-import queue still has pending updates');
     }
   } catch (error) {
-    // Check if pipeline is busy — defer rather than fail
     const isBusy = plugin.embedding_pipeline?.is_active() ||
       (error instanceof Error && error.message.includes('already processing'));
     if (isBusy) {
