+++ test/embed-job-queue.test.ts
@@ -4,7 +4,7 @@
  */
 
 import { describe, expect, it, vi } from 'vitest';
-import { EmbedJobQueue, type EmbedJob } from '../src/domain/embedding/queue/embed-job-queue';
+import { EmbedJobQueue, type EmbedJob } from '../src/domain/embedding/embed-job-queue';
 
 function makeJob(entityKey: string, contentHash: string = 'hash-' + entityKey): EmbedJob {
   return {
