+++ test/transformers-adapter.test.ts
@@ -4,7 +4,7 @@
  */
 
 import { describe, expect, it, vi } from 'vitest';
-import { TransformersEmbedAdapter, TRANSFORMERS_EMBED_MODELS } from '../src/ui/models/embed/adapters/transformers';
+import { TransformersEmbedAdapter, TRANSFORMERS_EMBED_MODELS } from '../src/ui/embed-adapters/transformers';
 
 function createAdapter(timeoutMs: number): TransformersEmbedAdapter {
   return new TransformersEmbedAdapter({
